# split-bill-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ActivityResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      detail:
        type: string
      group_id:
        type: string
      group_name:
        type: string
      id:
        type: string
      ref_id:
        type: string
      time_ago:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.ActivityType'
      user_avatar:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  models.ActivityType:
    enum:
    - bill_created
    - bill_deleted
    - bill_updated
    - member_joined
    - member_left
    - payment_sent
    - payment_confirmed
    - payment_rejected
    - group_created
    - settlement_created
    type: string
    x-enum-varnames:
    - ActivityBillCreated
    - ActivityBillDeleted
    - ActivityBillUpdated
    - ActivityMemberJoined
    - ActivityMemberLeft
    - ActivityPaymentSent
    - ActivityPaymentConfirmed
    - ActivityPaymentRejected
    - ActivityGroupCreated
    - ActivitySettlementCreated
...
  models.BankAccount:
    properties:
      account_name:
        type: string
      account_number:
        type: string
      bank_code:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      bank_accounts:
        items:
          $ref: '#/definitions/models.BankAccount'
        type: array
      display_name:
        maxLength: 50
        minLength: 2
        type: string
      preferred_payment:
        type: string
    type: object
  models.UserResponse:
    properties:
      avatar_url:
        type: string
      bank_accounts:
        items:
          $ref: '#/definitions/models.BankAccount'
        type: array
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      phone:
        type: string
      preferred_payment:
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@splitbill.app
    name: Split Bill Support
  description: API server for Split Bill application - manage groups, bills, transactions,
    OCR receipt scanning, and payments
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: 4.0.0
paths:
  /activities/me:
    get:
      description: Returns activities across all groups the authenticated user belongs
        to
      parameters:
      - description: Max number of activities (default 30)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ActivityResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user's activities
      tags:
      - Activities
  /auth/me:
    get:
      description: Returns the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /auth/profile:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile (display name, avatar,
        bank info)
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Auth
  /auth/verify-token:
    post:
      consumes:
      - application/json
      description: Verifies Firebase ID token and returns or creates the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Verify Firebase token
      tags:
      - Auth
...
  /groups/{id}/activities:
    get:
      description: Returns the activity feed for a specific group with optional limit
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Max number of activities (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ActivityResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get group activities
      tags:
      - Activities
...
securityDefinitions:
  BearerAuth:
    description: 'Firebase Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"